generate-package-tarball: reproducible node_modules tarballs

+++ util/generate-package-tarball.sh
@@ -7,36 +7,55 @@ if ! [ -f package.json ]; then
     exit 1
 fi
 
-rm -rf node_modules
+make() {
+    OUT_FILE="$1"
 
-pnpm install \
-    --frozen-lockfile \
-    --verify-store-integrity \
-    --shamefully-hoist
+    rm -rf node_modules
 
-TAR=gtar
-if ! command -v gtar 2>&1 >/dev/null
-then
-    TAR=tar
+    pnpm install \
+        --frozen-lockfile \
+        --verify-store-integrity \
+        --shamefully-hoist
+
+    TAR=gtar
+    if ! command -v gtar 2>&1 >/dev/null
+    then
+        TAR=tar
+    fi
+
+    pushd node_modules
+
+    rm -f .modules.yaml .pnpm-workspace-state.json
+
+    find . -print0 \
+        | sort -z \
+        | GZIP="-9n" $TAR -zcf "../$OUT_FILE.tar.gz" \
+            --sort=name \
+            --format=posix \
+            --numeric-owner \
+            --owner=0 \
+            --group=0 \
+            --pax-option='delete=atime,delete=ctime' \
+            --mode='go+u,go-w' \
+            --preserve-permissions \
+            --mtime='1970-01-01' \
+            --no-recursion \
+            --null \
+            --files-from -
+
+    popd
+}
+
+make a
+make b
+
+if ! cmp --silent -- a.tar.gz b.tar.gz; then
+    echo "something is broken, node_modules tarballs are not reproducible" >&2
+    sha256sum a.tar.gz b.tar.gz
+else
+    echo "ok, checksum: "
+    rm b.tar.gz
+    mv a.tar.gz node_modules.tar.gz
+    sha256sum node_modules.tar.gz
+    rm -rf node_modules
 fi
-
-pushd node_modules
-
-find . -print0 \
-    | sort -z \
-    | $TAR -zcf ../node_modules.tar.gz \
-        --format=posix \
-        --numeric-owner \
-        --owner=0 \
-        --group=0 \
-        --preserve-permissions \
-        --mtime='1970-01-01' \
-        --no-recursion \
-        --null \
-        --files-from -
-
-popd
-
-du -sh node_modules
-ls -lh node_modules.tar.gz
-rm -rf node_modules
